requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
claude-agent-sdk>=0.1.19
